refactor(ai-proxy): rename handler_mcp package to handler_mcp_server- Updated package name from handler_mcp to handler_mcp_server in handler.go - Updated import reference in provider.go to use new handler_mcp_server package
