+++ internal/apps/ai-proxy/handlers/handler_mcp_server/handler.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package handler_mcp
+package handler_mcp_server
 
 import (
 	"context"
